llcscraper/requirements.txt
Flask==3.0.0
APScheduler==3.10.4
keyring==24.3.0
python-dotenv==1.0.0
requests>=2.32.4
googlesearch-python==1.3.0
Jinja2==3.1.4
